+++ test/ripemd160.test.js
@@ -5,6 +5,13 @@ require('terst')
 describe('+ ripemd160(input)', function() {
   describe('> when input is a string', function() {
     it('should compute the ripemd160 hash', function() {
+      var input = "hello";
+      EQ (ripemd160(input).toString('hex'), "108f07b8382412612c048d07d13f814118445acd")
+    })
+  })
+
+  describe('> when input is a buffer', function() {
+    it('should compute the ripemd160 hash', function() {
       var input = new Buffer("hello");
       EQ (ripemd160(input).toString('hex'), "108f07b8382412612c048d07d13f814118445acd")
     })