+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ------------------
 * removed bower.json and component.json
 * changed 4 spacing to 2
-* returns `Buffer` type now, input must be Array, Uint8Array, or Buffer
+* returns `Buffer` type now, input must be Array, Uint8Array, Buffer, or string
 
 0.1.0 / 2013-11-20
 ------------------